수정 삭제 시 이미지 삭제 조건 수정
같은 이미지를 여러 글에서 사용하고 있을 경우를 생각해
이미지를 사용하는 글이 1개 일때만 삭제함
+ 메뉴바 스타일
+ 사진

+++ delete_article.php
@@ -28,9 +28,14 @@
 
 		//$imgurl = '"'.getcwd().$path.'"';
 		
-		$imgurl = substr($img["bgimg"], 2);		// 상대 경로를 사용해서 삭제함
+		$t_url = $db->quote($img["bgimg"]);
+		$anotherimgs = $db->query("SELECT num FROM $article_table WHERE bgimg=$t_url");
+		if($anotherimgs->rowCount() < 2)
+		{
+			$imgurl = substr($img["bgimg"], 2);		// 상대 경로를 사용해서 삭제함
 
-		unlink($imgurl);
+			unlink($imgurl);
+		}
 
 		$db->exec("DELETE FROM $article_table WHERE num=$num");
 		$db->exec("DELETE FROM $comment_table WHERE num=$num");
